refactor: move test class
